requirements.txt
streamlit
python-docx
latex2mathml
lxml